% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitSC.R
\name{LongitSC}
\alias{LongitSC}
\title{The function to fit a longitudinal bivariate binary model for semi-competing risks data using Psplines}
\usage{
LongitSC(longit.data, times = NULL, formula.NT, formula.T,
  formula.OR = NULL, data, epsOr = 10^(-10), knots = NULL,
  lambda = 0, init = NULL, maxit.optim = 50000)
}
\arguments{
\item{longit.data}{A data.frame or a list with columns named \code{risk.NT}, \code{risk.T},  \code{YNT}, \code{YT}. 
The function \code{TimesToLongit} can be used to obtain this representation of semicompeting risks time-to-event data.}

\item{times}{Vector of increasing times}

\item{formula.NT}{A formula of the form \code{YNT ~ x1 + x2} where \code{x1} and \code{x2} are covariates to be used for 
for the non terminal probaility sub-model.}

\item{formula.T}{A formula of the form \code{YT ~ x1 + x3} where \code{x1} and \code{x3} are covariates to be used for 
for the terminal probaility sub-model.}

\item{formula.OR}{A formula of the form \code{ ~ x1 + x4} where \code{x1} and \code{x4} are covariates to be used for 
for the odds ratio sub-model.}

\item{data}{A data.frame with the covariates specfied in \code{formula.NT}, \code{formula.T} and \code{formula.OR}.}

\item{epsOr}{How close it the OR allowed to be to one before assuming it equals to one. Default is \code{10^(-10)}}

\item{knots}{Number of knots for the Bsplines}

\item{lambda}{Penalization level. Either a vector of three values or a single number to be used for all three time-varying functions.}

\item{init}{Initial values for the parameters}

\item{maxit.optim}{For iternal use of \code{optim}. Default is 5000}
}
\value{
The function returns an object of class \code{LongitSC} including estimates and confidence intervals for 
the time-varying functions and coefficients.
}
\description{
The function implments the proposed methodlogy in Nevo et al. (2020+) for time-fixed covaraites under possible 
right-censoring and left truncation. Data should be first converted to longitudinal bivariate binary represenation. 
This could be done using \code{TimesToLongit}.
}
\note{
For time-varying covariates use \code{LongitSCtimeDep}. For unrestricted baseline functions 
(no Bsplines or penalization) use \code{LongitSCparamTimeDep}
 @author Daniel Nevo
}
