% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitSCparam.R
\name{LongitSCparam}
\alias{LongitSCparam}
\title{The function to fit a longitudinal bivariate binary model for semi-competing risks data with time-fixed covaraites
and using unrestricted time-varying functions.}
\usage{
LongitSCparam(longit.data, times = NULL, formula.NT, formula.T,
  formula.OR = NULL, data, epsOr = 10^(-10), knots = NULL,
  lambda = 0, init = NULL, maxit.optim = 50000)
}
\arguments{
\item{longit.data}{A data.frame or a list with columns named \code{risk.NT}, \code{risk.T},  \code{YNT}, \code{YT}. 
The function \code{TimesToLongit} can be used to obtain this representation of semicompeting risks time-to-event data.}

\item{times}{Vector of increasing times (for example, the interval partition points \eqn{\tau_1}... \eqn{\tau_K}).
This vector is used to construct the B-splines}

\item{formula.NT}{A formula of the form \code{YNT ~ x1 + x2} where \code{x1} and \code{x2} are covariates to be used for 
for the non terminal probability sub-model.}

\item{formula.T}{A formula of the form \code{YT ~ x1 + x3} where \code{x1} and \code{x3} are covariates to be used for 
for the terminal probability sub-model.}

\item{formula.OR}{A formula of the form \code{ ~ x1 + x4} where \code{x1} and \code{x4} are covariates to be used for 
for the odds ratio sub-model.}

\item{data}{A data.frame with the covariates specified in \code{formula.NT}, \code{formula.T} and \code{formula.OR}.}

\item{epsOr}{How close it the OR allowed to be to one before assuming it equals to one. Default is \code{10^(-10)}}

\item{knots}{Number of knots for the Bsplines}

\item{lambda}{Penalization level. Either a vector of three values or a single number to be used for all three time-varying 
functions.}

\item{init}{Initial values for the parameters.}

\item{maxit.optim}{For internal use of \code{optim}. Default is 50000}
}
\value{
The function returns an object of class \code{LongitSC} including estimates and confidence intervals for 
the time-varying functions and coefficients.
}
\description{
The function implements the proposed methodology in Nevo et al. (2020+) for time-fixed covariates under possible 
right-censoring and left truncation. Data should be first converted to longitudinal bivariate binary representation. 
This could be done using \code{TimesToLongit}. This function does not restricet the time-varying functions, and assume a
parameter for each interval. As a result, it contains more parameters and carry out parametric inference.
and implements penalized maximum likelihood to fit the model.
}
\note{
For time-varying covariates use \code{\link{LongitSCtimeDep}}. For unrestricted baseline functions 
(no B-splines or penalization) use \code{\link{LongitSCparamTimeDep}}.
 @author Daniel Nevo
}
