% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrData}
\alias{scrData}
\title{Ths scrData set from SemiCompRisks pacakge}
\format{
An object of class \code{data.frame} with 2000 rows and 8 columns.
}
\usage{
scrData
}
\description{
Ths scrData set from SemiCompRisks pacakge
}
\keyword{datasets}
