% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLongitDAta.R
\name{SimLongitData}
\alias{SimLongitData}
\title{The function that simulates data form longitudinal bivariate binary semicompeting risks data with baseline covarites}
\usage{
SimLongitData(n.sample, times = 1:100, beta.y, alpha.nt, alpha.t,
  alpha.or, beta.nt, beta.t, beta.or, X = NULL)
}
\arguments{
\item{n.sample}{Desired sample size.}

\item{times}{A vector of increasing times. Normally of equal length}

\item{beta.y}{True value for \eqn{\beta_y}.}

\item{alpha.nt}{True value for \eqn{\alpha_1(t)} for each \eqn{t} in \code{times}.}

\item{alpha.t}{True value for \eqn{\alpha_2(t)} for each \eqn{t} in \code{times}.}

\item{alpha.or}{True value for \eqn{\alpha_\theta(t)} for each \eqn{t} in \code{times}.}

\item{beta.nt}{True value for \eqn{\beta_1}.}

\item{beta.t}{True value for \eqn{\beta_2}.}

\item{beta.or}{True value for \eqn{\beta_\theta}.}

\item{X}{A matrix of time-fixed covariates to be used for simulating the data. Number of rows should be 
\code{n.sample} and 
number of columns should be equal to \code{length(beta.nt)}. If not specfied, \code{X} is simulated 
as iid Gaussian random variables.}
}
\value{
A list with the covariates used (\code{X}), at-risk indicators for each unit at each interval
 (\code{risk.NT} and \code{risk.T}) and outcome data
 at each interval (\code{YNT} and \code{YT}).
 @author Daniel Nevo
}
\description{
Given observed non-terminal and terminal event times, censoring indicators and possible left-truncation, 
this function returns the outcome data in the longitudinal bivariate binary representation according to given interval partition
 as proposed in Nevo et al.
}
