% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitSCparamTimeDep.R
\name{LongitSCparamTimeDep}
\alias{LongitSCparamTimeDep}
\title{The function to fit a longitudinal bivariate binary model for semi-competing risks data with time-depending 
covariates and using unrestricted (per-interval parameter) time-varying functions.}
\usage{
LongitSCparamTimeDep(
  times = NULL,
  data,
  formula.NT,
  formula.T,
  formula.OR = NULL,
  epsOR = 10^(-10),
  init = NULL,
  maxit.optim = 50000
)
}
\arguments{
\item{times}{Vector of increasing times (for example, the interval partition points \eqn{\tau_1,}..., \eqn{\tau_K}).
This vector is used to construct the B-splines}

\item{data}{A data.frame or a list with columns named \code{ID}, \code{TM}   \code{YNT}, \code{YT} as well as all covariate 
names used in \code{formula.NT}, \code{formula.T} and \code{formula.OR}. See details below. Other names can be used for
\code{YNT}, \code{YT}, but then their names should be given in the formulas below.}

\item{formula.NT}{A formula of the form \code{YNT ~ x1 + x2} where \code{x1} and \code{x2} are covariates to be used for 
the non terminal probability sub-model.}

\item{formula.T}{A formula of the form \code{YT ~ x1 + x3} where \code{x1} and \code{x3} are covariates to be used for 
the terminal probability sub-model.}

\item{formula.OR}{A formula of the form \code{ ~ x1 + x4} where \code{x1} and \code{x4} are covariates to be used for 
the odds ratio sub-model.}

\item{epsOR}{How close it the OR allowed to be to one before assuming it equals to one. Default is \code{10^(-10)}}

\item{init}{Initial values for the parameters.}

\item{maxit.optim}{For internal use of \code{optim}. Default is 50000}
}
\value{
The function returns an object of class \code{LongitSC} including estimates and confidence intervals for 
the time-varying functions and coefficients.
}
\description{
The function implements the proposed methodology in Nevo et al. (2020+) for time-depending covariates under
possible right censoring and left truncation. Data should be first converted to longitudinal bivariate binary representation,
similiar to the counting-process representation. See details below. This function does not restrict the time-varying functions and estimates
parameter for each interval. As a result, it contains more parameters and carry out parametric inference.
}
\details{
For \code{data}, the representation is similar to the counting process representation of time-to-event data.  
\code{ID} identify each person, where \code{TM} identifies the intervals in which this person is under followup.
\code{YNT} and \code{YT} indicate whether a non-terminal event and the terminal event occurred by the end of interval \code{TM}.
See examples in  \code{\link{SimLongitDataTimeDep}}.
}
\note{
For B-splines model for the time-varying function with penalized maximum likelihood use \code{\link{LongitSCtimeDep}}.
For time-fixed covariates use \code{\link{LongitSCtimeDep}}.
}
\examples{
\dontrun{
# Simulate semicompeting risks data
set.seed(314)
times <- seq(1,15,1)
alpha.nt <- LongitSemiComp:::logit(dchisq(times,3, ncp =5)/2 + 0.025)
alpha.t <- LongitSemiComp:::logit(times*(0.075/10)  - 0.0005*(times/20)^2  + 0.05)
alpha.or <- 0.15 - times/10 + 0.75*(times/10)^2 + 0.3*(times/20)^3
plot(x = times, y= exp(alpha.or))
plot(x = times, y= LongitSemiComp:::expit(alpha.nt))
plot(x = times, y= LongitSemiComp:::expit(alpha.t))
beta.nt <- log(c(0.7, 3))
beta.t <- log(c(0.5, 1))
beta.or <- log(c(1.4, 1))
beta.y <- log(1.4)
sim.data <- SimLongitDataTimeDep(n.sample = 1500, times = times,  beta.y,  cens.poten.rate = 0.5,
                                 alpha.nt, alpha.t, alpha.or, 
                                 beta.nt, beta.t, beta.or)
# Censoring rate
mean(sim.data$cens)
my.df <- sim.data$df.return
# Analysis
res <- LongitSCparamTimeDep(data = my.df, times = times,  formula.NT = YNT ~ X.1 + X.2, 
                            formula.T = YT  ~ X.1 + X.2, 
                            formula.OR = ~ X.1 + X.2)
 res
}

}
\author{
Daniel Nevo
}
