% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimesToLongit.R
\name{TimesToLongit}
\alias{TimesToLongit}
\title{Transfomring semicompeting risks outcome data to longitudinal bivariate binary representation}
\usage{
TimesToLongit(
  T1,
  T2,
  delta1,
  delta2,
  inter.vec,
  HandleCens = c("Right", "Left", "NN"),
  TruncData = F,
  TruncTime = NULL,
  HandleTrunc = c("Left", "Right", "NN")
)
}
\arguments{
\item{T1}{Observed non-terminal event time.}

\item{T2}{Observed terminal event time.}

\item{delta1}{Non-terminal event indicator (1: no-terminal event 0: censored)}

\item{delta2}{Terminal event indicator (1: no-terminal event 0: censored)}

\item{inter.vec}{Partition points creating the interval.}

\item{HandleCens}{How to handle within-interval right censoring (Section 3.2 of the paper). Right: anti-conservative
Left: conservative. NN: nearest neighbor.}

\item{TruncData}{Logical. Is the data include left truncation?}

\item{TruncTime}{Truncation time. Only used if TruncData==T}

\item{HandleTrunc}{How to handle within-interval left truncation (Section 3.2 of the paper). Right: conservative
Left: anti-conservative. NN: nearest neighbor.}
}
\value{
A list with at-risk indicators for each unit in each interval (risk.NT and risk.T) and outcome data
 at each interval (YNT and YT).
}
\description{
Given observed non-terminal and terminal event times, censoring indicators and possible left-truncation, 
this function returns the outcome data in the longitudinal bivariate binary representation according to given interval partition
 as proposed in Nevo et al.
}
\examples{
\dontrun{
data(scrData)
plot(scrData$time1, y = scrData$time2)
longit.data <- TimesToLongit(T1 = scrData$time1, T2 = scrData$time2, delta1 = scrData$event1, 
                             delta2 = scrData$event2, inter.vec = seq(0, 60, 5))
longit.data$YNT[1:3,]
longit.data$YT[1:3,]
longit.data$risk.NT[1:3,]
longit.data$risk.T[1:3,]
}
}
